fix(dns_resolver_test.py): update test to use float durations

+++ test/dns_resolver_test.py
@@ -5,7 +5,7 @@ from dnscrawler.logger import log
 import json
 import time
 if __name__ == "__main__":
-    start_time = int(time.time())
+    start_time = float(time.time())
     # resolver = DNSResolver([{'addr':"192.172.226.186",'port':1080}])
     resolver = DNSResolver(ipv4_only=True)
     # resolver = DNSResolver()
@@ -18,7 +18,7 @@ if __name__ == "__main__":
     domain_dict = resolver.get_domain_dict("amazon.com")
     # Early ns records
     # domain_dict = resolver.get_domain_dict("aridns.net.au")
-    finish_time = int(time.time())
+    finish_time = float(time.time())
     duration = finish_time - start_time
     print(json.dumps(domain_dict))
     print(f"Duration:{duration}")
